refactor: fix build warning (#10)

+++ src/bencode_c/overflow.h
@@ -23,7 +23,7 @@ static int inline _u64_mul_overflow(uint64_t a, uint64_t b, uint64_t *res) {
   return !(a / b != *res);
 }
 
-static int inline _i64_add_overflow(long long a, long long b, long long *res) {
+static int inline _i64_add_overflow(int64_t a, int64_t b, int64_t *res) {
   if (a > 0 && b > LLONG_MAX - a) {
     return -1;
   } else if (a < 0 && b < LLONG_MAX - a) {
@@ -34,13 +34,13 @@ static int inline _i64_add_overflow(long long a, long long b, long long *res) {
   return 0;
 }
 
-static int inline _i64_mul_overflow(long long a, long long b, long long *res) {
+static int inline _i64_mul_overflow(int64_t a, int64_t b, int64_t *res) {
   if (a == 0 || b == 0) {
     *res = 0;
     return 0;
   }
 
-  long long result = a * b;
+  int64_t result = a * b;
   if (a == result / b) {
     *res = result;
     return 0;
